implement NEC tx

+++ panasonic_tx.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-void send_panasonic_ircode(uint32_t device, uint16_t code);
+void send_panasonic_ircode(uint16_t device, uint16_t code);
 void panasonic_register_functions(void (*pwm_on_function)(void),
                                   void (*pwm_off_function)(void),
                                   void (*delay_us_function)(unsigned int),
